feat: moving the period calculation to a helper function with tests

+++ libs/backend/ranking/src/services/calculation/calculation.service.ts
@@ -9,6 +9,7 @@ import moment, { Moment } from 'moment';
 import { Op, Transaction } from 'sequelize';
 import { PlaceService } from '../place';
 import { PointsService } from '../points';
+import { getRankingPeriods } from '@badman/utils';
 
 @Injectable()
 export class CalculationService {
@@ -17,57 +18,72 @@ export class CalculationService {
   constructor(
     private pointsService: PointsService,
     private placeService: PlaceService,
-  ) {}
-
-  public async simulation(
-    systemId: string,
-    fromDate?: Date | string,
-    toDate?: Date | string,
-    periods?: number,
-    transaction?: Transaction,
   ) {
-    const system = await RankingSystem.findByPk(systemId, {
-      include: [{ model: RankingGroup }],
-    });
+    require('events').EventEmitter.defaultMaxListeners = Infinity;
+  }
+
+  public async updateRanking(
+    system: string | RankingSystem,
+    args: {
+      fromDate?: Date | string;
+      toDate?: Date | string;
+      periods?: number;
+      recalculatePoints?: boolean;
+      transaction?: Transaction;
+    } = {
+      recalculatePoints: false,
+    },
+  ) {
+    if (typeof system === 'string') {
+      system = (await RankingSystem.findByPk(system, {
+        include: [{ model: RankingGroup }],
+      })) as RankingSystem;
+    }
+
+    const { fromDate, toDate, periods, recalculatePoints, transaction } = args;
 
     if (!system) {
-      throw new NotFoundException(`${RankingSystem.name}: ${systemId}`);
+      throw new NotFoundException(`${RankingSystem.name} not found`);
     }
     system.runCurrently = true;
     await system.save();
 
-
     try {
-
       this.logger.log(`Simulation for ${system.name}`);
 
       // if no periods are defined, calulate up untill last update interval
       const toDateM = moment(toDate);
-      const fromDateM = moment(fromDate);
+      const fromDateM = moment(
+        fromDate || system.updateIntervalAmountLastUpdate,
+      );
       if ((periods ?? 0) > 0) {
         for (let i = 0; i < (periods ?? 0); i++) {
           fromDateM.subtract(
-            system.caluclationIntervalAmount,
+            system.calculationIntervalAmount,
             system.calculationIntervalUnit,
           );
         }
       }
 
-      const updates = this._getUpdateIntervals(system, fromDateM, toDateM);
+      const updates = getRankingPeriods(system, fromDateM, toDateM);
 
-      // I know t
-      const minUpdatePlace = moment(updates[0][0]);
-      const minUpdatePoints = moment(updates[0][0]).subtract(
+      if (updates.length === 0) {
+        this.logger.log(`Simulation finished ${system.name}`);
+        return;
+      }
+
+      const minUpdatePlace = moment(updates[0].date);
+      const minUpdatePoints = moment(updates[0].date).subtract(
         system.periodAmount,
         system.periodUnit,
       );
-      const maxUpdate = moment(updates[updates.length - 1][0]);
+      const maxUpdate = moment(updates[updates.length - 1].date);
 
       this.logger.log(
         `Simulation for ${system.name} has ${
           updates.length
         } point updates planned, including ${
-          updates.filter((u) => u[1]).length
+          updates.filter((u) => u.updatePossible).length
         } ranking updates, between ${minUpdatePlace?.format(
           'YYYY-MM-DD',
         )} and ${maxUpdate?.format('YYYY-MM-DD')}
@@ -84,46 +100,54 @@ export class CalculationService {
         transaction,
       });
 
-      await RankingPoint.destroy({
-        where: {
-          systemId: system.id,
-          rankingDate: {
-            [Op.between]: [minUpdatePoints.toDate(), maxUpdate.toDate()],
-          },
-        },
-        transaction,
-      });
-
-      for (let period = 0; period < (system.periodAmount ?? 0); period++) {
-        this.logger.debug(
-          ` points for date: ${moment(minUpdatePoints).format(
+      if (recalculatePoints) {
+        this.logger.verbose(
+          `Recalculate points for ${
+            system.name
+          }, between ${minUpdatePoints?.format(
             'YYYY-MM-DD',
-          )}, ${period} / ${system.periodAmount}}`,
+          )} and ${maxUpdate?.format('YYYY-MM-DD')}`,
         );
-
-        await this.pointsService.createRankingPointsForPeriod({
-          system,
-          calcDate: minUpdatePoints.toDate(),
-          options: {
-            transaction,
+        await RankingPoint.destroy({
+          where: {
+            systemId: system.id,
+            rankingDate: {
+              [Op.between]: [minUpdatePoints.toDate(), maxUpdate.toDate()],
+            },
           },
+          transaction,
         });
 
-        minUpdatePoints.add(1, system.periodUnit);
+        for (let period = 0; period < (system.periodAmount ?? 0); period++) {
+          this.logger.debug(
+            ` points for date: ${moment(minUpdatePoints).format(
+              'YYYY-MM-DD',
+            )}, ${period} / ${system.periodAmount}}`,
+          );
 
+          await this.pointsService.createRankingPointsForPeriod({
+            system,
+            calcDate: minUpdatePoints.toDate(),
+            options: {
+              transaction,
+            },
+          });
+
+          minUpdatePoints.add(1, system.periodUnit);
+        }
       }
 
-      for (const [index, [updateDate, isUpdateNeeded]] of updates.entries()) {
+      for (const [index, { date, updatePossible }] of updates.entries()) {
         this.logger.debug(
-          `${moment(updateDate).format(
+          `${moment(date).format(
             'YYYY-MM-DD',
-          )}, ${isUpdateNeeded}, ${index} / ${updates.length}}`,
+          )}, ${updatePossible}, ${index} / ${updates.length}}`,
         );
 
         const startUpdate = moment();
         await this.pointsService.createRankingPointsForPeriod({
           system,
-          calcDate: updateDate,
+          calcDate: date,
           options: {
             transaction,
           },
@@ -131,10 +155,10 @@ export class CalculationService {
 
         await this.placeService.createUpdateRanking({
           system,
-          calcDate: updateDate,
+          calcDate: date,
           options: {
             transaction,
-            updateRanking: isUpdateNeeded,
+            updateRanking: updatePossible,
           },
         });
 
@@ -162,57 +186,5 @@ export class CalculationService {
       system.runCurrently = false;
       await system.save();
     }
-  }
-
-  private _getUpdateIntervals(system: RankingSystem, from: Moment, to: Moment) {
-    let limit = 100;
-    const lastUpdate = moment(system.updateIntervalAmountLastUpdate);
-    const updates: [Date, boolean][] = [];
-
-    if (!system.updateIntervalAmount || !system.updateIntervalUnit) {
-      throw new Error('No update interval defined');
-    }
-
-    while (lastUpdate.isAfter(from) || limit <= 0) {
-      lastUpdate
-        .subtract(system.updateIntervalAmount, system.updateIntervalUnit)
-        .startOf('month')
-        .add(6, 'day')
-        .startOf('isoWeek');
-      limit--;
-    }
-
-    while (from.isBefore(to)) {
-      const diff = lastUpdate.diff(from, system.updateIntervalUnit) * -1;
-      const isUpdateNeeded = diff >= system.updateIntervalAmount;
-
-      if (isUpdateNeeded) {
-        lastUpdate
-          .add(system.updateIntervalAmount, system.updateIntervalUnit)
-          .startOf('month')
-          .add(6, 'day')
-          .startOf('isoWeek');
-      }
-
-      updates.push([from.toDate(), isUpdateNeeded]);
-
-      from.add(
-        system.caluclationIntervalAmount,
-        system.calculationIntervalUnit,
-      );
-    }
-
-    // sort updates by date
-    updates.sort((a, b) => {
-      if (a[0].getTime() > b[0].getTime()) {
-        return 1;
-      }
-      if (a[0].getTime() < b[0].getTime()) {
-        return -1;
-      }
-      return 0;
-    });
-
-    return updates;
   }
 }
